scripts/pyproject.toml
[project]
name = "pubmed-oa-corpus-scripts"
version = "0.1.0"
description = "Scripts to download PubMed Central open access papers for RAG evaluation"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.83",
    "sickle>=0.7.0",
    "lxml>=5.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "bandit>=1.7.0",
    "types-tqdm>=4.66.0",
    "types-pyyaml>=6.0.0",
    "lxml-stubs>=0.5.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit
]
ignore = [
    "S311",   # random is fine for jitter
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
# Bio.Entrez and sickle are untyped libraries
module = ["Bio.*", "sickle.*"]
ignore_missing_imports = true
ignore_errors = true

[tool.bandit]
skips = ["B311"]  # random is fine for jitter
